Move everserver
