+++ pyproject.toml
@@ -77,7 +77,7 @@ dependencies = [
 ert = "ert.__main__:main"
 "fm_dispatch.py" = "_ert.forward_model_runner.fm_dispatch:main"
 everest = "everest.bin.main:start_everest"
-everserver = "everest.detached.jobs.everserver:main"
+everserver = "everest.detached.everserver:main"
 recovery_factor = "everest.jobs.scripts.recovery_factor:main"
 render = "everest.jobs.scripts.render:main"
 wdcompl = "everest.jobs.scripts.wdcompl:main"
